requirements.txt
aiohttp-jinja2==1.4.2
pydantic==1.8.2
pytest-aiohttp==0.3.0